# res/openapi/huki-ebook-openapi.generated.json
{
  "components": {
    "schemas": {
...
      "commerce_CancelOrderDto": {
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "commerce_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "community_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "identity_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-auth": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-refresh": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "shipping_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "https://huki.ebook.com",
      "name": "HUKI Support",
      "url": "support@huki.ebook.com"
    },
    "description": "\n# HUKI EBOOK Platform API\n\nWelcome to the HUKI EBOOK API documentation. This API provides endpoints for managing books, orders, payments, and more.\n\n## Base URL\n`http://localhost:3000/api/v1`\n\n## Authentication\nMost endpoints require JWT Bearer token authentication.\n1. Register/Login to get access token\n2. Include header: `Authorization: Bearer <token>`\n\n## Services\n| Service | Port | Description |\n|---------|------|-------------|\n| Identity | 3001 | Auth, Users, Sessions |\n| Business | 3002 | Business, Stores, Members |\n| Commerce | 3003 | Books, Cart, Orders, Payments |\n| Shipping | 3004 | Shipments, Addresses |\n| Community | 3005 | Forum, Chat, Reviews |\n| Promotion | 3007 | Vouchers, Banners |\n\n## Common Headers\n| Header | Required | Description |\n|--------|----------|-------------|\n| Content-Type | Yes | application/json |\n| Authorization | No | Bearer token |\n| Idempotency-Key | For checkout | Unique key for idempotent operations |\n\n## Error Codes\n- 400: Bad Request\n- 401: Unauthorized\n- 403: Forbidden\n- 404: Not Found\n- 409: Conflict\n- 500: Server Error\n\n---\n*Last updated: 2026-08-24*\n    ",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://huki.ebook.com/terms",
...
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "description": "Returns a paginated list of orders for the authenticated user.",
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "PENDING_PAYMENT",
                "PROCESSING",
                "SHIPPING",
                "COMPLETED",
                "CANCELLED",
                "PARTIALLY_CANCELLED",
                "REFUNDED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of orders"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List user orders",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "description": "Returns detailed information about a specific order.",
        "operationId": "OrdersController_detail",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "403": {
            "description": "Order does not belong to user"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order details",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "post": {
        "description": "Cancels an order. Only possible before payment or shipment.",
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commerce_CancelOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order cancelled successfully"
          },
          "400": {
            "description": "Order cannot be cancelled (already shipped/paid)"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "403": {
            "description": "Order does not belong to user"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/history": {
      "get": {
        "description": "Returns the complete status change history for an order.",
        "operationId": "OrdersController_history",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order status history"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "403": {
            "description": "Order does not belong to user"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order status history",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/tracking": {
      "get": {
        "description": "Returns real-time tracking information for an order.",
        "operationId": "OrdersController_tracking",
        "parameters": [
          {
            "description": "Order ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order tracking information"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "403": {
            "description": "Order does not belong to user"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Track order status",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/payments/orders/{orderId}": {
      "get": {
        "operationId": "PaymentsController_status",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payment and refund status for an order",
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/v1/seller/orders/{id}": {
      "get": {
        "operationId": "SellerOrdersController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get seller order details",
        "tags": [
          "Seller orders"
        ]
      }
    },
    "/api/v1/seller/orders/{id}/cancel": {
      "patch": {
        "operationId": "SellerOrdersController_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commerce_CancelOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an order",
        "tags": [
          "Seller orders"
        ]
      }
    },
...
  },
  "servers": [
...
}